testing statements

+++ FtcRobotController/src/main/java/com/qualcomm/ftcrobotcontroller/opmodes/caftc/PlayAutonomous.java
@@ -45,8 +45,8 @@ public class PlayAutonomous extends BasicAutonomous{
     long lTargetTime;
     long rTargetTime;
 
-    int lEncoderValue;
-    int rEncoderValue;
+    long lEncoderValue;
+    long rEncoderValue;
 
     @Override
     public void init(){
